20.3.CSV Export{Hiding the Extra}
Hiding the Extra
No worries, we just need to hide that link manually... and we already have a filter to do this! Open EasyAdminExtension and filterActions(). Now, just unset ['export']. That looks a little crazy, so I'll add a comment: "This action is rendered manually"
Try it! Yes! We have the export button on top... but not on each row. This is a tricky - but valid - use-case for custom actions.