21.1.Tweaking the Form Layout{Form Layout Config Customizations}
We've talked a lot about customizing the forms... which mostly means using the config to change the field types or adding a custom form theme to control how the individual fields look.
But, what about the form layout? Like, what if I wanted to put the email & full name fields in a section that floats to the left, and these other two fields in a section that floats to the right?
Well... that's not really part of the form component: we usually do this by adding markup to our template. But of course, the template lives inside EasyAdminBundle!
In turns out, there are two great ways to control your form's layout. First..., well, you could just override the template and go crazy! For example, inside the bundle, find new.html.twig. It has a block called entity_form. And, to render the form, it just calls the form() function. This means that there's no form layout at all by default: it just barfs out all the fields.
But... that's awesome! Because we could override this template, replace just the entity_form block, and go bonkers by rendering the form however we need. And since we can override each template on an entity-by-entity basis... well... suddenly it's super easy to customize the exact form layout for each section.

Form Layout Config Customizations
Phew! But... there is an even easier way that works for about 90% of the use-cases. And that is... of course... with configuration. EasyAdminBundle comes with a bunch of different ways to add dividers, sections and groups inside the form.
So, let's do it! Start with User. Let's reorganize things: put fullName on top, then add a new type called divider. Put avatarUri after the divider, then another divider, email, divider, isScientist and universityName.
On a technical level... this is kind of geeky cool: divider is a fake form field! I mean, in the bundle itself, it is literally a form field. And you can see a few others, like section and group. We'll use all three.
Try out the page! Hello divider! It's nothing too fancy, but it's nice.