21.3.Tweaking the Form Layout{Reorganizing all Fields under form}
Reorganizing all Fields under form
The last organizational trick is the group, which gives you mad control over the width and float of a bunch of fields.
To see an example, go up to the Genus form.
And first, remember how we organized most fields under the form key... but then tweaked a few final things under new and edit? Well, when EasyAdminBundle reads this, all of the fields under form are added first... and then any extra fields under new or edit are added. That means, in edit, our slug field is printed last on the form. And... there's not really a good way to control that. This gets even a little bit more problematic when you want to organize fields into sections or groups. How could you organize the slug field into the same section as name? Right now, you can't!
For that reason, it's best to configure all of your fields under form. Then, use new and edit only to remove fields you don't want. Copy the slug field and remove edit entirely. Then, under form, paste this near the top.
To keep slug off of the new form, just add -slug.
The end result is the same, but with complete control over the field order.