21.4.Tweaking the Form Layout{Adding Form Groups}
Adding Form Groups
Ok, back to adding groups. First, move id and slug to the end of the form. Then, on top, add a new group: type: group, css_class: 'col-sm-6', label: 'Basic Information'.
You can picture what this is doing: adding a div with col-sm-6, putting a header inside of it, and then printing any fields below that, but in the div.
And that's huge! Because thanks to the col-sm-6 CSS class, we can really start organizing how things look.
Move funFact and isPublished a bit further down. Then, after subFamily, add a section labeled Optional.
Yep, you can totally mix-and-match groups and sections.
At this point, funFact and isPublished will still be in the group, but they'll also be in a section within that group. And since genusScientists is pretty big, let's put that in their own group with css_class: col-sm-6 and label: Studied by...
Finally, at the bottom, add one more group. I'll use the expanded format this time: css_class: col-sm-6 and label: Identification. And yep, groups can have icon and help keys.
Phew! While I did this, I added some line breaks just so that this all looks a bit more clear: here's one group, here's a second group, the last group is at the bottom.
But what does it actually look like? Let's find out! Refresh!
Oh, this feels good. The "Basic Information" group is on the left with the "Optional section" at the bottom. The other two groups float to the right.
Now, sometimes, you might want to force the Identification group to go onto its own line. Basically, you want to add a CSS clear after the first two groups.
To do that, on the group, add a special CSS class, called new-row.
And now it floats to the next line. So, groups are a really, really neat way to control how things are rendered. It adds some nice markup, and we can add whatever classes we need. So, there's not much you can't do.