22.2.Dashboard & Menu Customizations{Adding a Dashboard}
Adding a Dashboard
If you downloaded the course code, you should have a tutorial/ directory. Inside, it has an AdminController with a dashboardAction(). Copy that. Then, in src/Controller/EasyAdmin, open our AdminController and paste it there.
Thanks to the prefix on the route import:
config/routes/easy_admin.yaml
easy_admin_bundle:
    prefix: /easyadmin
This creates a new /easyadmin/dashboard route named admin_dashboard. Oh, I'm missing my use statement for @Route. I'll re-type that and hit enter so that it auto-completes the use statement on top.
Perfect!
This renders a template, which I will also grab from the tutorial/ directory. Paste that in templates/easy_admin.
At this point... the page should work. Cool... but how can I tell EasyAdminBundle to show this page when we go to the admin section's homepage? Right now, if you go directly to /easyadmin, it will take you to whatever the first-defined entity section is... so Genus.