23.1.Customizing the Menu
What else can we do with the menu? Well, of course, we can expand the simple entity items to get more control. To expand it, add entity: User. Now we can go crazy with label: Users.
Remember, you can also specify the label under the User entity itself. If we added a label key here, it would apply to the menu, the header to that section and a few other places. But under menu, it just changes the menu text.
Do the same thing for GenusNote, with label: Notes, and also for sub families: entity: SubFamily, label: Sub-Families.
At this point, it should be no surprise that we can control the icon for each menu. Like, icon: user, icon: sticky-note and icon: ''
Before configuring anything, each item has a folder icon. With empty quotes, even that icon is gone.