23.3.Customizing the Menu{Adding Menu Separators & Sub-Menus}
Or, you can go a step further and create a sub-menu. Change this new menu item to use the expanded format. Then, add a children key. Indent all the other links so that they live under this.
And just to make it even nicer, add a separator called Related.
Try it! Try it! Nice! The Genus menu expands to show the sub-items and the Related separator.