23.5.Customizing the Menu{Custom Links}
On that note, one of the other query parameters is action, which we can also set to anything. Copy this entire new menu link and - at the top of children - paste it. This time, let's link to the show page of 1 specific genus... our favorite "Pet genus". To do that, set action to show and id to some id in the database, like 2.
This isn't anything special, we're just taking advantage of how the query parameters work in EasyAdminBundle.
And while we're here, it might also be nice to add a link to the front-end of our app. This is also nothing special: add a new link that points to the app_genus_list route called "Open front-end".
Refresh! And try that link. Nice!