23.6.Customizing the Menu{External Links}
External Links
In addition to routes, if you want, you can just link to external URLs. Go to the bottom of the list... and make sure we're at the root level. Add a new section called "Important stuff" with icon: explanation and a children key.
I'll paste a couple of very important external links for silly kittens and wet cats.
Yep, instead of entity or route keys, you can skip all of that and just add url. And of course, you can set the target on any item.